little-isesues

+++ deliverboh/database/migrations/2021_11_23_170428_create_dishes_table.php
@@ -18,7 +18,7 @@ class CreateDishesTable extends Migration
             $table->foreignId('user_id')->constrained();
             $table->string('name', 100);
             $table->string('ingredients')->nullable();
-            $table->float('price', 7,2);
+            $table->decimal('price', 7,2);
             $table->text('description');
             $table->boolean('visibility');
             $table->string('image')->nullable();
